+++ app/src/main/res/layout/esterilizacion_list_fragment.xml
@@ -2,11 +2,9 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:background="#4D80EA82"
+    android:background="@color/adopcion"
     android:layout_marginBottom="3dp"
     android:layout_marginHorizontal="5dp">
