+++ app/src/main/res/xml/files.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<paths>
+    <external-path name = "crime_photos"
+        path="/" />
+</paths>
